Separates Tests From Components
