+++ src/component/tests/FoodDetailsInfo.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-import FoodDetailsInfo from './FoodDetailsInfo';
+import FoodDetailsInfo from '../FoodDetailsInfo';
 
 describe('Food Details Information Component Tests', () => {
   it('matches the Food Details Information component snapshot', () => {
