+++ src/component/tests/Foods.test.js
@@ -3,8 +3,8 @@ import renderer from 'react-test-renderer';
 import { BrowserRouter as Router } from 'react-router-dom';
 import { render, screen } from '@testing-library/react';
 import { Provider } from 'react-redux';
-import Foods from './Foods';
-import store from '../redux/store';
+import Foods from '../Foods';
+import store from '../../redux/store';
 
 describe('Foods Component Tests', () => {
   it('should display the name of food', () => {
