+++ src/component/tests/Loading.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-import Loading from './Loading';
+import Loading from '../Loading';
 
 describe('Loading Component Tests', () => {
   it('matches the Loading component snapshot', () => {
